requirements.txt
backtrader==1.9.78.123
tushare==1.4.25
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=6.0
matplotlib>=3.5.0
jinja2>=3.0.0
pytest>=7.0